Sync kombu_reconnect_delay from Oslo
Back ported from Icehouse.
When reconnecting to a RabbitMQ cluster with mirrored queues in use, the attempt to release the connection can hang "indefinitely" somewhere deep down
in Kombu.  Blocking the thread for a bit prior to release seems to kludge around the problem where it is otherwise reproduceable.
The value 5.0 fits for low perfomance environments as well.

Implements: rally-user-story US6888
Upstream-Review: https://review.openstack.org/#/c/95489/
Not-in-upstream: true